Add unstable feature handling to compile on beta channel
It sounds like u32:BITS is still in flux, at least as far as where it
will end up being stabilized. In order to compile on beta I've hardcoded
it in as 32, which is fine for nightly as well since the number of bits
for a u32 will always be 32.

It also looks like the extern crate test isn't used so I've removed it.